Filter Memento by Image Classification (#320)
* image_labels must be str for get route

Using a list in the filter params model, when the model is part of a
route will require the list to be sent in request body. This is not
allowed for a GET endpoint.

* Adding image_labels to filter logic in memento query

* Endpoint to get Image Labels from user Mementos

* image labels hook

* fix image labels hook

* Image Label Dropdown

* image labels always stale

* Formatted Label

* add image_label to apply filters

* syncing with backend

* Adding comments

* linting

* Fixing incorrect docstring

* Image Label Filter with one query

* Using inner join with get memento query

Supabase uses left join by default. When filtering the image labels, if
an "images" does not have an image with the corresponding label it will
be an empty list instead of removing the memento from the response.

* Default Selection that clears Image Label Filter

* Formatted Input Value after Filters Applied

When filters are applied for chosen image label, the non-formatted value
was shown when filter sheet is reopened. e.g., choose Chocolate Sauce,
apply filters, open filters, image_label is chocolate_sauce.

* Fix image label field alignment

* Refetch image labels when filter sheet opens

* only accept one image_label in filter query

squash

update memento query

* Format image label on backend send both to frontend

* sync frontend

* Update filter query image_labels field

* Use ImageLabelResponse fields instead of formatting

* Format image label on ImageWithUrl

* lint