Auto derive collection metadata (#318)
* Add mementos button redirects to a page to select keepsakes

* Test lower opacity upon click

* Opacity of memento changes when clicked

* Mark ID when memento is clicked

* Pass selected mementos back to create collection page

* Fixed back arrow

* Add mementos to create collection query

* Wait for mementos to load on selection screen

* Add selected memento cards to form

* Horizontal scrollview for selected mementos

* Cleaned up code

* Documentation

* Change color of select mementos button; green checkmark on selected mementos

* Undo changes to MementoCard

* Moving select_mementos screen into collections group

On Android, router was unable to find the screen under
(select_mementos)/select_mementos.tsx

* Update metadata aggregation to accept Photo or Memento

Need to aggregate metadata from either Photo or Mementos. Create
agnostic Metadata type that can be extracted from Photo or Memento.

* Selected Mementos using Controller

By using a controller here, we have finer control over updating the form
data and auto-deriving location and date.

* Using MementoWithImages instead of Memento

MementosWithImages is type received from backend and has coordinates of
type Coordinates.

* Edit collection mementos fixed

* Pre-select mementos selected on the collection form last

* sync api

* CollectionForm Auto-Derives Date and Location

When adding Mementos to a Collection, the Date and Location fields are
automatically derived from the aggregated date and coordinates from the
Mementos. These fields are derived in the same way the Date and Location
fields for Mementos are derived from the added Photos.

Adding Checkbox for Autofill Modal

* Adding Checkbox for Autofill modal

* Autofill Modal First Draft

Missing value in Checkbox props, not used but gives warning.

* Modal with CollectionForm

* Updating for Add Memento Changes

* Fixing pnpm-lock

* Using toISODateString to drop Timezone

* Rearranging in CollectionForm

* Autofillmodal only if back from select_mementos

* Resolving diagnostics

* Divider in autofill modal

* Updating comments

* Fixing for checks

* simplify update_collection_and_mementos

* Only renders checkbox if derived metadata found

* fix bug #319

---------

Co-authored-by: Mahmud Jamal <mahmudjamal014@gmail.com>
Co-authored-by: owencooke <ocooke@ualberta.ca>